+++ src/engine/Graphics/Context/ExcaliburGraphicsContext2DCanvas.ts
@@ -152,6 +152,14 @@ export class ExcaliburGraphicsContext2DCanvas implements ExcaliburGraphicsContex
     dwidth?: number,
     dheight?: number
   ): void {
+    if (swidth === 0 || sheight === 0) {
+      return; // zero dimension dest exit early
+    } else if (dwidth === 0 || dheight === 0) {
+      return; // zero dimension dest exit early
+    } else if (image.width === 0 || image.height === 0) {
+      return; // zero dimension source exit early
+    }
+
     this.__ctx.globalAlpha = this.opacity;
     const args = [image, sx, sy, swidth, sheight, dx, dy, dwidth, dheight]
       .filter((a) => a !== undefined)